Fix : 좋아요 관련 이슈 해결 #182

+++ MSA/Recruitment/src/main/java/com/springcooler/sgma/recruitmentboardlike/command/application/service/RecruitmentBoardLikeServiceImpl.java
@@ -61,25 +61,33 @@ public class RecruitmentBoardLikeServiceImpl implements RecruitmentBoardLikeComm
     public RecruitmentBoardLike deleteLike(Long recruitmentBoardId, Long userId) {
         Optional<RecruitmentBoard> optionalRecruitmentBoard = recruitmentBoardRepository.findById(recruitmentBoardId);
         List<RecruitmentBoardLikeDTO> recruitmentBoardLikes = recruitmentBoardLikeService.findAllRecruitmentBoardLike();
-        System.out.println("===========================");
-        for (RecruitmentBoardLikeDTO like : recruitmentBoardLikes) {
-            if (like.getRecruitmentBoardId().equals(recruitmentBoardId) && like.getUserId().equals(userId)) {
-                RecruitmentBoard recruitmentBoard = optionalRecruitmentBoard.get();
-                System.out.println("recruitmentBoard"+recruitmentBoard);
+        System.out.println("123===========================");
+        System.out.println(recruitmentBoardLikes);
+        if(recruitmentBoardLikes.size()==0) {
+            throw new CommonException(ErrorCode.NOT_FOUND_LIKE);
+        }
+        else {
+            for (RecruitmentBoardLikeDTO like : recruitmentBoardLikes) {
+                System.out.println(recruitmentBoardId);
+                System.out.println(like.getRecruitmentBoardId());
+                if (like.getRecruitmentBoardId().equals(recruitmentBoardId) && like.getUserId().equals(userId)) {
+                    RecruitmentBoard recruitmentBoard = optionalRecruitmentBoard.get();
+                    System.out.println("recruitmentBoard" + recruitmentBoard);
 
-                RecruitmentBoard updatedRecruitmentBoard = recruitmentBoard.toBuilder()
-                        .likes(recruitmentBoard.getLikes()-1)
-                        .build();
-                System.out.println("updatedRecruitmentBoard"+updatedRecruitmentBoard);
-                RecruitmentBoardLike recruitmentLike = recruitmentBoardLikeRepository.findByRecruitmentBoardIdAndUserId(recruitmentBoardId, userId);
-                recruitmentBoardLikeRepository.delete(recruitmentLike);
-                System.out.println("좋아요가 취소되었습니다.");
+                    RecruitmentBoard updatedRecruitmentBoard = recruitmentBoard.toBuilder()
+                            .likes(recruitmentBoard.getLikes() - 1)
+                            .build();
+                    System.out.println("updatedRecruitmentBoard" + updatedRecruitmentBoard);
+                    RecruitmentBoardLike recruitmentLike = recruitmentBoardLikeRepository.findByRecruitmentBoardIdAndUserId(recruitmentBoardId, userId);
+                    recruitmentBoardLikeRepository.delete(recruitmentLike);
+                    System.out.println("좋아요가 취소되었습니다.");
 
-                recruitmentBoardRepository.save(updatedRecruitmentBoard);
-                return null;
-            }
-            else{
-                throw new CommonException(ErrorCode.NOT_FOUND_LIKE);
+                    recruitmentBoardRepository.save(updatedRecruitmentBoard);
+                    return null;
+                } else {
+                    System.out.println("123456");
+                    throw new CommonException(ErrorCode.NOT_FOUND_LIKE);
+                }
             }
         }
         return null;
